// src/models/models.ts
export interface IPokemon {
    abilities: Ability[];
    base_experience: number;
    forms: Form[];
    game_indices: GameIndice[];
    height: number;
    held_items: any[];
    id: number;
    is_default: boolean;
    location_area_encounters: string;
    moves: Move[];
    name: string;
    order: number;
    past_types: any[];
    species: Species;
    sprites: Sprites;
    stats: Stat[];
    types: Type[];
    weight: number;
    isFavourite?: boolean
}

export interface IActions {
    [Actions.View]?: string
    [Actions.Favourites]?: string
}

export interface IFilter {
    key: string,
    type: FilterTypes,
    label?: string,
    placeholder?: string
    span?: number
}

export interface DataType {
    id: number;
    name: string
    height: number;
}

export interface IFiltersSettings {
    title: string,
    filters: IFilter[]
}

export interface ITableSettings {
    title: string,
    columns: Array<ColumnType<IPokemon>>
    actionOptions: Array<string>
}

export interface IPageSettings {
    filtersSettings: IFiltersSettings;
    tableSettings: ITableSettings
}